monitoring/requirements.txt
grpcio==1.35.0
opencensus-ext-stackdriver==0.7.4
opencensus==0.7.12
absl-py



